requirements.txt
transformers==4.35.2
torch==2.1.1
pyperclip==1.8.2
Pillow==10.1.0
